Walking locomotion changes
walking_locomotion now overrides get_observations, this is because the standing policy and walking policy need different observations.

Locomotion_gym_env has rolled back to when it was operational only for the standing policy.

Changed default INIT_MOTOR_ANGLE to be the default_dof provided by the simulation. Now whenever the robot starts it's default angle it will start in the default dof_position