Walking locomotion and Robot changes
Moved standing function from robot reset to walking_locomotion.py's init function. This will put the robot in default_dof position (standing position) before starting the walking policy. This will allow it to be positioned correctly before running the policy.

Changed the step function motor_control_mode to always be set to position.

Created an observation for when the robot is on the rack. Artificially set velocities to make the robot believe it is on the ground moving.